[APPC-1006] Solve issues caused by wrongly moved file
