+++ app/src/main/java/com/asfoundation/wallet/ui/BaseActivity.java
@@ -1,4 +1,4 @@
-package com.asfoundation.wallet.ui.balance;
+package com.asfoundation.wallet.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
