+++ app/src/main/java/com/asfoundation/wallet/ui/Erc681Receiver.java
@@ -9,7 +9,6 @@ import com.asf.wallet.R;
 import com.asfoundation.wallet.entity.TransactionBuilder;
 import com.asfoundation.wallet.interact.FindDefaultWalletInteract;
 import com.asfoundation.wallet.interact.PaymentReceiverInteract;
-import com.asfoundation.wallet.ui.balance.BaseActivity;
 import com.asfoundation.wallet.ui.iab.IabActivity;
 import com.asfoundation.wallet.ui.iab.InAppPurchaseInteractor;
 import com.asfoundation.wallet.util.TransferParser;
