+++ app/src/main/java/com/asfoundation/wallet/ui/MyAddressActivity.java
@@ -16,7 +16,6 @@ import androidx.lifecycle.ViewModelProviders;
 import com.asf.wallet.R;
 import com.asfoundation.wallet.entity.NetworkInfo;
 import com.asfoundation.wallet.entity.Wallet;
-import com.asfoundation.wallet.ui.balance.BaseActivity;
 import com.asfoundation.wallet.viewmodel.MyAddressViewModel;
 import com.asfoundation.wallet.viewmodel.MyAddressViewModelFactory;
 import com.google.zxing.BarcodeFormat;
