+++ app/src/main/java/com/asfoundation/wallet/ui/WalletsActivity.java
@@ -17,7 +17,6 @@ import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import com.asf.wallet.R;
 import com.asfoundation.wallet.entity.ErrorEnvelope;
 import com.asfoundation.wallet.entity.Wallet;
-import com.asfoundation.wallet.ui.balance.BaseActivity;
 import com.asfoundation.wallet.ui.widget.adapter.WalletsAdapter;
 import com.asfoundation.wallet.util.KeyboardUtils;
 import com.asfoundation.wallet.viewmodel.WalletsViewModel;
