+++ app/src/main/java/com/asfoundation/wallet/ui/gamification/RewardsLevelActivity.kt
@@ -5,7 +5,7 @@ import android.view.Menu
 import android.view.MenuItem
 import com.asf.wallet.R
 import com.asfoundation.wallet.router.TransactionsRouter
-import com.asfoundation.wallet.ui.balance.BaseActivity
+import com.asfoundation.wallet.ui.BaseActivity
 
 class RewardsLevelActivity : BaseActivity(), GamificationView {
