+++ app/src/main/java/com/asfoundation/wallet/ui/transact/TransferActivity.kt
@@ -8,7 +8,7 @@ import android.os.Bundle
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 import com.asf.wallet.R
-import com.asfoundation.wallet.ui.balance.BaseActivity
+import com.asfoundation.wallet.ui.BaseActivity
 import com.asfoundation.wallet.ui.barcode.BarcodeCaptureActivity
 import java.math.BigDecimal
